Implement concurrent catalog/thread posts processing.
Upon loading catalog/thread N posts around the last viewed position (not currently implemented by
will be in the future) are parsed eagerly. Right after that a parallel task is invoked which starts
parsing the rest of the posts. A small notification is displayed at the bottom of catalog/thread which
indicates that posts are being processed. This makes catalog/thread loading way faster than in the original
KurobaEx which parses posts all posts before displaying them.