Fix the damn thing that was causing lots of unnecessary recompositions tanking the performance.
The problem was in code that is calling postsScreenViewModel.onPostListBuilt(), namely we were
re-executing the checks to figure out whether we have a visible PostCell on every recomposition
which, for some reason, was causing more recompositions to happen (basically infinite). I have
no idea why it's causing more recompositions because nothing in LazyListState is modified.