Fix incorrect layout mode used for a split second after changing phone orientation.
We now reset the currentOrientation once we detect that the phone is rotated which resets
the mainUiLayoutMode to null and other stuff that depends on it so that while the orientation
is being changed nothing is composed, thus no state is used incorrectly.

This fixes a bug where the toolbar search query is not restored sometimes after phone rotation.