Fix post getting out of sort order in sticky rolling threads.
The cause of this bug was us never processing the deleted posts that we store in the ChanCache. Once a
post is deleted we don't receive it from the server anymore so we don't add it into the list of posts
that we then submit to parsing. However during post parsing we still can add that post into the result
post list if there is a quote to that post. Then when merging posts in the PostsState we always append
posts that are not in postIndexes at the end of the list and that's how the list get reordered.

Also added some dev flavor checks that will crash the app if we detect that the post list is unordered.