Fix thread bookmarking not working as expected when clicking the Bookmark icon.
The problem was in the function addOrRemoveBookmark() getting called twice in a row because we had
2 subscribers to the same "IconClicks" state from two different places due to changes in NavigationRouter
code that caused two of the same controllers (for example two CatalogControllers) being added
into the list that was used for transition between two toolbars.

Implement goToPrev()/goToNext() actions of the mini media player.