Fix update checks.
When doing force check, compare version from the server and current app version (from the BuildConfig).
This is needed to make sure we always show the "New app version" notification when we are not using
the latest version.

Otherwise when doing auto-check on app start, compare version from the server and the app version
checked the last time. This is needed to only show the "New app version" notification once when
doing auto-update checks and to not spam the user with notifications.