Do not hold a persistent reference to an activity inside of ComposeScreens.
This fixes the memory leak where upon screen rotation the activity is referenced from
all the screens which are stored in the MainActivityViewModel which is a 100% memory leak.

To fix that the activity is nulled out when activity.onDestroy() callback is called.
When activity.onCreate() is called all the screens in the screen tree are receiving the
new activity reference which they pass to all child screens etc.