Fix toolbar animations not working as expected.
The problem was caused a division by zero inside of pagerState.currentPageOffsetFraction
which then returned a NaN float which when used by Modifier.graphicsLayer { ... }
breaks it somewhere internally probably on the RenderNode level. Basically it makes it
behave incorrectly: in case of alpha it make it constantly slightly transparent, in case
of translation it places everything with some additional offset.