Merge KPNC into KurobaExLite.
Unfortunately it's impossible to keep push notifications as a separate application because, apparently, it's impossible to wake up another application with enabled network by sending a broadcast receiver when the device is in deep sleep mode (doze mode). When FCM message comes from Firebase it only wakes up the app that is listening for FCM messages. Other apps are still "sleeping". So when you send a broadcast receiver to another app it will wake it up but there will be no network connection so you won't be able to go a check for new posts on the server.